refactor: Standardize render/math helper APIs on float and remove redundant double bridges (#147)
* remove double overload for CTRRectangle

* remove double to float casting

* update makergba to use float

* remove double overload

* remove double overload for rgbacolor

* remove double cast

* remove unnecessary double cast

* remove cast

* remove unnecessary cast mess

* remove double cast

* remove double cast

* unnecessary casts, gone

+++ CutTheRope/Framework/Visual/DrawHelper.cs
@@ -152,7 +152,7 @@ namespace CutTheRope.Framework.Visual
 
         public static void CalcCircle(float x, float y, float radius, int vertexCount, float[] glVertices)
         {
-            float num = (float)(6.283185307179586 / vertexCount);
+            float num = 6.283185307179586f / vertexCount;
             float num2 = 0f;
             for (int i = 0; i < vertexCount; i++)
             {
